finish

+++ app/Http/Controllers/ProjectController.php
@@ -4,9 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AddUsersToProjectRequest;
 use App\Http\Requests\StoreProjectRequest;
+use App\Http\Requests\UpdateProjectRequest;
 use App\Models\Project;
+use App\Models\User;
 use App\Services\ProjectService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
@@ -24,11 +27,26 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        //
+        $user = Auth::user();
+        $user = User::findOrFail($user->id);
+        $projects = Project::with('users')->get();
+
+        // Ensure that there is an authenticated user
+        if (!$user || (!$user->is_admin)) {
+            abort(response()->json([
+                'error' => 'You are not authorized to perform this action.',
+            ], 403));
+        }
+        return response()->json($projects, 200);
     }
 
+
+
     /**
-     * Store a newly created resource in storage.
+     * create new task.
+     * @param StoreProjectRequest $request
+     * @return \Illuminate\HTTP\JsonResponse
+
      */
     public function store(StoreProjectRequest $request)
     {
@@ -46,13 +64,30 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
-        //
+        $user = Auth::user();
+        $user = User::findOrFail($user->id);
+        $project = Project::findOrFail($project->id);
+        // Check if the user is associated with the project
+        $userProjectRelation = $project->users()->where('user_id', $user->id)->first()?->pivot->role;
+        if (!$user->is_admin && !$userProjectRelation) {
+            abort(response()->json([
+                'message' => 'User is not associated with this project.',
+            ], 404));
+        }
+        // Ensure that there is an authenticated user
+        if (!$user || (!$user->is_admin && !$userProjectRelation)) {
+            abort(response()->json([
+                'error' => 'You are not authorized to perform this action.',
+            ], 403));
+        }
+        $project = $project->load('users');
+        return response()->json($project, 200);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Project $project)
+    public function update(UpdateProjectRequest $request, Project $project)
     {
         $validatedData = $request->validated();
         $project = $this->projectService->updateProject($project, $validatedData);
@@ -63,11 +98,23 @@ class ProjectController extends Controller
         }
     }
 
+
     /**
-     * Remove the specified resource from storage.
+     * Remove project from storage.
+     * @param Project $project
+     * @return \Illuminate\HTTP\JsonResponse
+
      */
     public function destroy(Project $project)
     {
+        $user = Auth::user();
+        // Ensure that there is an authenticated user
+        if (!$user || (!$user->is_admin && $this->user()->id !== $user->id)) {
+            abort(response()->json([
+                'error' => 'You are not authorized to perform this action.',
+            ], 403));
+        }
+
         $project = $this->projectService->deleteProject($project);
         if (isset($project['error'])) {
             return response(['error' => $project['error'], 'message' => $project['message']], $project['status']);
@@ -75,18 +122,15 @@ class ProjectController extends Controller
             return response(['status' => $project['status'], 'message' => $project['message']], $project['status']);
         }
     }
+    /**
+     * add Users To Project by admin.
+     * @param AddUsersToProjectRequest $request
+     * @param  $projectId
+     * @return \Illuminate\HTTP\JsonResponse
 
+     */
     public function addUsersToProject(AddUsersToProjectRequest $request, $projectId)
     {
-        // {
-        //     "users": [
-        //       { "id": 1, "role": "admin" },
-        //       { "id": 2, "role": "member" },
-        //       { "id": 3, "role": "manager" }
-        //     ]
-        //   }
-
-        // Validate the request data
         $validatedData = $request->validated();
 
         $project = $this->projectService->addUsersToProjectPivot($projectId, $validatedData);
@@ -94,6 +138,57 @@ class ProjectController extends Controller
             return response(['error' => $project['error'], 'message' => $project['message']], $project['status']);
         } else {
             return response(['status' => $project['status'], 'message' => $project['message'], 'project' => $projectId, 'project_users' => $project['project_users']], $project['status']);
+        }
+    }
+    /**
+     * set the start time of work on project.
+     * @param  $projectId
+     * @param  $userId
+     * @return \Illuminate\HTTP\JsonResponse
+
+     */
+    public function startProject($projectId, $userId)
+    {
+        $user = Auth::user();
+        // Ensure that there is an authenticated user
+        if (!$user) {
+            abort(response()->json([
+                'error' => 'You are not authorized to perform this action.',
+            ], 403));
+        }
+        $project = $this->projectService->startTime($projectId, $userId);
+        if (isset($project['error'])) {
+            return response(['error' => $project['error'], 'message' => $project['message']], $project['status']);
+        } else {
+            return response(['status' => $project['status'], 'message' => $project['message'], 'task_start_time' => $project['start_time']], $project['status']);
+        }
+    }
+    /**
+     * set the end time of work on project.
+     * @param  $projectId
+     * @param  $userId
+     * @return \Illuminate\HTTP\JsonResponse
+
+     */
+    public function endProject($projectId, $userId)
+    {
+        $user = Auth::user();
+        // Ensure that there is an authenticated user
+        if (!$user) {
+            abort(response()->json([
+                'error' => 'You are not authorized to perform this action.',
+            ], 403));
+        }
+        $project = $this->projectService->endTime($projectId, $userId);
+        if (isset($task['error'])) {
+            return response(['error' => $project['error'], 'message' => $project['message']], $project['status']);
+        } else {
+            return response()->json([
+                'message' => $project['message'],
+                'session_hours' => $project['session_hours'],
+                'userTaskInfo' => $project['userTaskInfo'],
+                'end_time' => $project['end_time'],
+            ], $project['status']);
         }
     }
 }
